pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Web page analyzer for SEO acceptance"
authors = ["Viktor Shcherbakov <evil.logitech@gmail.com>"]
license = "Freeware"
readme = "README.md"
packages = [{include = "page_analyzer"}]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "License :: Freeware",
    "Natural Language :: English"
]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.3"
gunicorn = "^21.2.0"

# Poetry pre-1.2.x style, Understood by Poetry 1.0-1.2
psycopg2-binary = "^2.9.7"
validators = "^0.21.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
